install: add support for --dry-run and --help

+++ install.sh
@@ -8,7 +8,32 @@ then
   exit 1
 fi
 
-set -ex
+usage () {
+  echo "Usage: $(basename "$0") [-v|--verbose] [--dry-run] [PREFIX]"
+}
+
+while [[ $# -gt 0 ]]; do
+  case "$1" in
+    --help)
+      usage
+      exit 0
+      ;;
+    --verbose|-v)
+      set -x
+      shift
+      ;;
+    --dry-run)
+      DRY_RUN=1
+      shift
+      ;;
+    *)
+      usage
+      exit 1
+      ;;
+  esac
+done
+
+set -e
 
 CARGO_USER=${SUDO_USER:-root}
 
@@ -16,12 +41,14 @@ PREFIX=${1:-/usr/local}
 TMP_INSTALL_DIR="$SCRIPT_DIR/target/install"
 sudo -u "$CARGO_USER" -i cargo install --force --path "$SCRIPT_DIR" --root "$TMP_INSTALL_DIR" --no-track
 
-install -D -t "$PREFIX"/bin/ "$TMP_INSTALL_DIR"/bin/udev-hid-bpf
-
 sed -e "s|/usr/local|$PREFIX|" 99-hid-bpf.rules > target/bpf/99-hid-bpf.rules
 
-install -D -t /lib/firmware/hid/bpf target/bpf/*.bpf.o
-install -D -m 644 -t /etc/udev/rules.d target/bpf/99-hid-bpf.rules
-install -D -m 644 -t /etc/udev/hwdb.d target/bpf/99-hid-bpf.hwdb
-udevadm control --reload
-systemd-hwdb update
+if [[ -z "$DRY_RUN" ]];
+then
+  install -D -t "$PREFIX"/bin/ "$TMP_INSTALL_DIR"/bin/udev-hid-bpf
+  install -D -t /lib/firmware/hid/bpf target/bpf/*.bpf.o
+  install -D -m 644 -t /etc/udev/rules.d target/bpf/99-hid-bpf.rules
+  install -D -m 644 -t /etc/udev/hwdb.d target/bpf/99-hid-bpf.hwdb
+  udevadm control --reload
+  systemd-hwdb update
+fi
